// frontend/src/lib/api/types.ts
export interface OrderItem {
  id: number;
  product_id: number | null;
  product_name: string;
  unit_price: number;
  formatted_unit_price: string;
  quantity: number;
  subtotal: number;
  formatted_subtotal: string;
}

export interface Shipment {
  id: number;
  courier: string;
  service: string;
  tracking_number: string | null;
  status: string;
}

export interface Payment {
  id: number;
  payment_type: string | null;
  status: string;
  amount: number;
}

export interface Order {
  id: number;
  user_id: number;
  status: string;
  subtotal: number;
  formatted_subtotal: string;
  shipping_cost: number;
  formatted_shipping_cost: string;
  total: number;
  formatted_total: string;
  shipping_courier: string;
  shipping_service: string;
  shipping_etd?: string | null;
  shipping_address: {
    name: string;
    phone: string;
    province: string;
    city: string;
    district: string;
    postal_code: string;
    address: string;
  };
  items: OrderItem[];
  shipment?: Shipment | null;
  payment?: Payment | null;
  created_at: string;
  updated_at: string;
}